Refactor Snackbar and Clipboard
