+++ static/js/fileIndex.js
@@ -1,4 +1,7 @@
 import { deleteFile } from "./controllers/delete.js";
+import { copyToClipboard } from "./lib/copyToClipboard.js";
+import { makeShortLink } from "./lib/links.js";
+import { createSnackbar } from "./lib/snackbar.js";
 
 const errorContainer = document.getElementById("error");
 
@@ -28,59 +31,17 @@ document.querySelectorAll('[pico-purpose="delete"]').forEach((deleteBtn) => {
 document.querySelector("#error .delete").addEventListener("click", () => {
   hideElement(errorContainer);
 });
+
 document.querySelectorAll('[pico-purpose="copy"]').forEach((copyBtn) => {
   copyBtn.addEventListener("click", () => {
     const picoId = copyBtn.getAttribute("pico-entry-id");
-    const shortLink = `${window.location.origin}/!${picoId}`;
+    const shortLink = makeShortLink(picoId);
 
     copyToClipboard(shortLink)
-      .then(() => createSnackbar())
+      .then(() => createSnackbar("Link was copied!"))
       .catch((error) => {
         document.getElementById("error-message").innerText = error;
         showElement(errorContainer);
       });
   });
 });
-
-/**
- * @param {string} textToCopy
- * @returns Promise
- */
-function copyToClipboard(textToCopy) {
-  // navigator clipboard api needs a secure context (https)
-  if (navigator.clipboard && window.isSecureContext) {
-    // navigator clipboard api method'
-    return navigator.clipboard.writeText(textToCopy);
-  } else {
-    // text area method
-    const textArea = document.createElement("textarea");
-    textArea.value = textToCopy;
-    // make the textarea out of viewport
-    textArea.style.position = "fixed";
-    textArea.style.left = "-999999px";
-    textArea.style.top = "-999999px";
-    document.body.appendChild(textArea);
-    textArea.focus();
-    textArea.select();
-    return new Promise((res, rej) => {
-      // here the magic happens
-      document.execCommand("copy") ? res() : rej();
-      textArea.remove();
-    });
-  }
-}
-
-function createSnackbar() {
-  const snackbarContainer = document.getElementById("snackbar-container");
-  const el = document.createElement("div");
-
-  el.classList.add("snackbar");
-  el.innerHTML = "Link was copied!";
-
-  snackbarContainer.append(el);
-  el.classList.add("show");
-
-  setTimeout(() => {
-    el.remove();
-  }, 3000);
-}
