+++ templates/file-index.html
@@ -1,5 +1,6 @@
 {{ define "style-tags" }}
 <link rel="stylesheet" type="text/css" href="/css/file-index.css" />
+<link rel="stylesheet" type="text/css" href="/css/lib/snackbar.css" />
 {{ end }}
 
 {{ define "script-tags" }}
@@ -59,8 +60,6 @@
     <div id="error-message" class="message-body">Placeholder error.</div>
   </article>
 </div>
-
-<div id="snackbar-container"></div>
 
 {{ end }}
